:new: :memo: meta: Add issue templates
This commit adds new issue templates for the Inventory Hugo theme.
Originally, I disabled issues on this repository to keep everything
together in a single repository. However, now that there are multiple
downstreams of this theme, it makes sense to use the issue tracker of
this repository to collect feedback about the design, features, and
UI/UX of the Hugo theme, separate from site-specific content.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>