:telephone_receiver: layouts: Add RSVP form for events
This commit adds a new layout for an RSVP form using the same Formspree
layout used for the contact page.

In the future, there is likely a better way to unify these as a common
layout in the event other form types are needed, but for now, this meets
the need for the Drones 4 SDG Toolkit webinar on 28 October 2021.

Related to unicef/drone-4sdgtoolkit#13.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>