:telephone_receiver: layouts: Add RSVP form for events (#26)
* :telephone_receiver: layouts: Add RSVP form for events

This commit adds a new layout for an RSVP form using the same Formspree
layout used for the contact page.

In the future, there is likely a better way to unify these as a common
layout in the event other form types are needed, but for now, this meets
the need for the Drones 4 SDG Toolkit webinar on 28 October 2021.

Related to unicef/drone-4sdgtoolkit#13.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :art: layouts: Update section-title to text-dark class

Addresses @Idadelveloper's feedback in unicef/drone-4sdgtoolkit#15. This
uses a darker color for the section title on the contact form and RSVP
form pages. It contrasts better with the lighter background color.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>