:globe_with_meridians: i18n: Bundle localization into theme
This commit takes the existing localization files from the example site
and embeds them into a directory where Hugo will look for localizations.
These can be overridden by creating a file of the same name in the Hugo
site, but would require copying all existing translations from the
upstream-provided file. Updates and changes would be manually managed.

This encourages upstreaming of localization efforts and also avoids
snags when a new site is launched with this theme, but the parts of the
theme expecting a `i18n/` configuration render empty.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>