:new: :sparkles: UI: Add edit button to all content pages (closes #50)
This commit adds an "Edit this Page" button to appear on all content
pages, if not hidden by a boolean configuration option in post front
matter. This makes it easier to find the source content and to propose
edits through the GitHub (or other git forge) web interface.

Authored by @Neha9849. Closes #50.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>