:new: :sparkles: UI: Add edit button to all content pages (closes #50) (#110)
This commit adds an "Edit this Page" button to appear on all content
pages, if not hidden by a boolean configuration option in post front
matter. This makes it easier to find the source content and to propose
edits through the GitHub (or another git forge) web interface.

Authored by @Neha9849. Closes #50.

* :globe_with_meridians: Add internationalization

This commit leverages the Hugo internationalization component and adds
the string as a translatable string. An English and French example are
provided to demonstrate.

ref: https://gohugo.io/content-management/multilingual/

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

Co-authored-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>