:wrench: config: Improve theme color contrast (#116)
I have checked out all the websites which are using this theme and I found out that except SustainOSS Design & UX knowledgebase all others are using the same color palette. So, I thought of adding the same color palette to the example site, so that it will be easier to inspect design issues and fix them.