:technologist: .github: Add user story issue template form
This commit adds a new GitHub Issues template, using the new forms
feature. The new template is modeled for user stories, which we already
use in project management planning for internships on this project.

Additionally, this also adds pointer links to the other toolkits that
use this theme, in order to help redirect a potential contributor to the
best place to leave their feedback.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>