:technologist: meta: Update Contributor Covenant (`v2.0` -> `v2.1`)
A minor change to the Code of Conduct. The language now adheres to the
latest version of the Contributor Covenant. Additionally, the format was
changed from AsciiDoc to Markdown so that GitHub's "Community Standards"
report card will detect that a Code of Conduct exists on this repository
(currently, it thinks it is missing).

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>