:sparkles: layouts: Add redirect layout and docs (closes #69) (#134)
This commit adds a new type of layout, a redirect, to the theme. A
redirect layout is similar to the existing Hugo aliases, except instead
of being limited to internal redirect (i.e. redirecting one page to
another page in the same site), we can now redirect a user to a page
hosted externally on another site.

The motivation for this is so that we can ensure that user experience is
not disrupted when content is moved from one site to somewhere else. In
our case, we want to move content about data science into the UNICEF
Data Science & A.I. Toolkit and some of our developer-focused content
into the Software Development Toolkit. This new feature enables that
transition to happen smoothly for anyone who has an old URL bookmarked
or linked somewhere else. See unicef/inventory#128 for more context.

Closes unicef/inventory-hugo-theme#69.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

:globe_with_meridians: i18n: Add localization support for redirect page

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>