:new: :memo: Add contributing category & key stakeholders (closes #127) (#137)
This commit adds a new "Contributing" category to the site and
initializes it with a "Key stakeholders" page. The stakeholders page is
part of a user story and defines the primary users for the Open Source
Inventory toolkit.

This page is part of a gradual effort of converting the example site for
the UNICEF Inventory theme into more comprehensive and useful
documentation. Eventually, more content may be migrated to live in this
category as a common reference for several toolkits.

Closes #127. Part of #89.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>