:lipstick: layouts(navigation): Fix responsive navbar clipping
This commit improves the overall responsiveness of the top navigation
bar. Previously, the text of the site would clip off-screen and disrupt
the user experience. This fixes the navigation bar so it appears better
on smaller screen sizes.

Additionally, support for the alternate logo was removed since it was
made redundant in this change, so the feature is removed from the code
and no longer exposed in the example site configuration.

Fixes #143.