:globe_with_meridians: Update translations from French to Spanish
To help demonstrate the multi-lingual features of this theme, we will
need to rely on translating some or all of our content. The more
translated content we have, the more comprehensive it will be for the
user to change languages and see the feature in action.

In a practical sense, there is not a strong French knowledge to keep
translating content into French. I am comfortably fluent in Spanish, and
while my translations are not perfect, I can better catch machine
translation errors in Spanish than French. So, in order to help me
continue to localize content in the new example site structure, this
changes the supported localization from French to Spanish.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>