:building_construction: exampleSite: Initialize new content structure (closes #89) (#148)
This change initializes the new structure of the UNICEF Inventory theme
example site, making it a more comprehensive tool for documenting the
theme itself.

The Pull Request is divided into different commits for easier review. A
summary of changes:

1. Translated content is changed from French to Spanish.
2. The _Billing & Pricing_ category was removed.
3. The _About_ category was added.
4. The config file includes minor brand and identity changes to reflect
    the purpose of the example site.

Closes #89.

* :globe_with_meridians: Update translations from French to Spanish

To help demonstrate the multi-lingual features of this theme, we will
need to rely on translating some or all of our content. The more
translated content we have, the more comprehensive it will be for the
user to change languages and see the feature in action.

In a practical sense, there is not a strong French knowledge to keep
translating content into French. I am comfortably fluent in Spanish, and
while my translations are not perfect, I can better catch machine
translation errors in Spanish than French. So, in order to help me
continue to localize content in the new example site structure, this
changes the supported localization from French to Spanish.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :boom: Purge Billing & Pricing category

Not applicable to our site or its use.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :new: :memo: Create new About category

This will describe the theme and how it is made.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* Update brand and identity in config

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

* :globe_with_meridians: Localize site footer

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>