:building_construction: exampleSite: Document key features (closes #119) (#149)
This change is a large reworking of content, structure, and organization.
Several unique features of the UNICEF Inventory theme are given a
feature profile to show how they work and how to use them. This makes
features easier to discover and also makes them searchable from the site
homepage.

Additionally, a "lorem ipsum" page is included in this change. It groups
several page elements together in a single page, to help a webmaster
evaluate the theme's appearance and for developers to test for regressions
when introducing new front-end changes.

Closes #119.

---

* :new: :memo: Create "Brand & identity" feature profile

Converts existing content in the Install category about customization as
a feature profile, where it better belongs.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :new: :technologist: Create a feature profile template post

For quicker creation of feature templates in the future.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :boom: Remove duplicate pages with feature descriptions

There are five pages in the Features category with the same content.
This commit removes four of them. This is de-duplication before the
existing content is rewritten into better indexed feature descriptions.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :building_construction: Adopt a structure better for i18n

If more localizations are added in the future, the directory will become
difficult to navigate with so many files in a single directory, even if
the number of pages are fewer.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :new: :memo: Move Author Byline from Install to Features as a profile

Moves @Neha9849's documentation into a feature profile in the new
category.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :memo: Add Tab Tables feature profile

Converted from the example page. This emphasizes the feature better, but
the purpose of the large example page as helpful for testing should
remain. To be revisited.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :memo: Add Admonitions feature profile

Calling this out as a specific feature because it was something that
came up in conversation with a downstream user, and it later became a
research question. Documenting and calling it out improves visibility of
the feature.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :memo: Add Code Highlighting feature profile

Another one that was a research question during Ida's internship and
could have been found more quickly with better documentation.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :new: :content: Add example page with several elements together

This creates a page that brings together several elements into a single
article. This is mainly for demonstrative and testing purposes.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :memo: Use a PDF with less legal ambiguity

Not sure where this PDF came from or where it was sourced. Using the PDF
print-out of www.unicef.org is a safer route by any means.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>

---

* :memo: Add description to Features category

Get rid of generic lorem ipsum text, and use something helpful. Also
adds a Spanish translation.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>