:boom: layouts: Remove unused single page template
The single page template is removed so it can be easier to work. The
option to embed the whole DPG Indicators layout into the layout itself
is not possible; a layout is limited to one `{{ define "" }}` statement
while a partial can contain many.

Later, this should be simplified so the reusable partials are cycled
among the various layouts so there is more code sharing, and less
duplication.

CC: @Neha9849

ref: https://gohugo.io/templates/single-page-templates/

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>