:new: :memo: exampleSite: Add forking and reuse docs (closes #118)
This commit creates new documentation about how to fork and reuse the
UNICEF Inventory theme in new projects, whether for a new Hugo website
or as a UNICEF Toolkit. This replaces the content formerly found in the
`/content/install/` directory.

Additionally, weights were assigned to all categories so that the
categories appear on the home page in a specific order. The weights
start from `20` because eventually, the `/content/about/` section will
use a weight of `10`, putting it before other categories.

Also, some Spanish content was added to demonstrate the multilingual
features of the site, but more work should be done to translate the
content to provide a meaningful example of this feature in practice.

Closes #118.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>