:bug: layouts(navigation): Remove extraneous HTML, simplify CSS classes
This commit removes the added `<p>` tag, which added extra padding and
spacing between the logo and site title. It takes the `navbar-brand`
class and applies it to the whole `<div>` while moving the `text-center`
class into the site title, so the spacing appears correctly.

This also removes competing CSS classes, simplifying debugging later.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>