:bug: layouts(index): Show "team" type posts on index page
This commit shows the "team" template type on the index page of the
site. The use case for this change is so that the total cohorts view
still appears when navigating to `/inventory/cohorts/` and it also
appears as its own category on the site homepage alongside other topics.

This contributes to unblocking unicef/inventory#136.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>