:art: :lipstick: layouts(index): Reduce padding around index text (#178)
This commit moves the index page text (if provided in the Hugo site)
into the `<section>` tag for the main topics. It uses a Bootstrap
container inside of the `<section>` and therefore exists inside of the
styling rules for the section instead of working against its styling.

This reduces the large amount of whitespace that exists between the
search bar section of the site and the topics. It is easier to realize
there is more content.

Related to unicef/inventory-hugo-theme#154.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>