:art: assets: Comment and organize CSS stylesheet
This commit is a complete reorganization of the custom CSS stylesheet
used in the UNICEF Inventory theme. Most CSS rules are now grouped
together in like sections with other rules. It is easier to read the
file and also understand which rules are associated to which part of the
theme. This is an effort to reduce technical debt and improve ease of
maintenance for future maintainers.

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>