:new: :memo: features: Add Team Profiles doc (closes #170)
This commit creates a new doc for the Team Profiles feature. This gives
information about the motivation for the feature, an example profile,
and a screenshot from the UNICEF Open Source Inventory. This continues
the ongoing effort to make the example site a useful documentation
resource about the theme and how to use it.

Closes unicef/inventory-hugo-theme#170.

CC: @siddharthvipul

Signed-off-by: Justin W. Flory (he/him) [UNICEF Innovation] <jflory@unicef.org>